更名 node-ajax.js 到 ajax.js 且不再包含在浏览器版本中；剥离 ajax.js 对全局变量 AV 的依赖
