+++ lib/av.js
@@ -12,6 +12,7 @@ var AV = module.exports = {};
 AV._ = require('underscore');
 AV.VERSION = require('./version');
 AV.Promise = require('./promise');
+AV.XMLHttpRequest = require('./browserify-wrapper/xmlhttprequest').XMLHttpRequest;
 AV.localStorage = require('./localstorage');
 
 // 以下模块为了兼容原有代码，使用这种加载方式。
@@ -33,8 +34,10 @@ require('./status')(AV);
 require('./search')(AV);
 require('./insight')(AV);
 
-// Node.js platform specific
-require('./browserify-wrapper/node-ajax')(AV);
+// Node.js platform specific patches
+if (AV._isNode) {
+  AV._ajax = require('./browserify-wrapper/ajax');
+}
 
 // Backward compatibility
 AV.AV = AV;
