+++ package.json
@@ -40,6 +40,7 @@
   },
   "browser": {
     "react-native": false,
+    "./lib/browserify-wrapper/ajax.js": false,
     "./lib/browserify-wrapper/upload.js": "./lib/browserify-wrapper/upload-browser.js",
     "./lib/browserify-wrapper/xmlhttprequest.js": "./lib/browserify-wrapper/xmlhttprequest-browser.js",
     "./lib/browserify-wrapper/localStorage.js": "./lib/browserify-wrapper/localstorage-browser.js",
